+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Additional labels for pre-release and build metadata are available as extensions
 
 ## [2.4.0-dev]
 1. Fixed pandas<3.0.0 and numpy<2.3.1 dependencies, ([#458](https://github.com/brightwind-dev/brightwind/issues/458))
-2. Added `lapse_temperature` to output a temperature value for any height by lapsing a known reference temperature value at a different measurement height ([[#530](https://github.com/brightwind-dev/brightwind/issues/530)]) 
+2. Added `extrapolate_temperature` to output a temperature value for any height by lapsing a known reference temperature value at a different measurement height ([[#530](https://github.com/brightwind-dev/brightwind/issues/530)]) 
 
 ## [2.3.0]
 This update brings a comprehensive set of **bug fixes** and **enhancements** across the Brightwind library.